feat: Update mediaquery size
