+++ utils/mediaquery.js
@@ -1,6 +1,6 @@
 import facepaint from 'facepaint';
 
-const breakpoints = [480, 768];
+const breakpoints = [768, 900];
 
 const mediaquery = facepaint(
   breakpoints.map((breakpoint) => `@media (min-width: ${breakpoint}px)`)
